refactor: vitest config cli update optional
