+++ packages/cli/src/helpers/index.ts
@@ -6,7 +6,9 @@ export * from "./getPackageRunner";
 export * from "./installWithRunner";
 export * from "./isModuleType";
 export * from "./makeLogMessage";
-export * from "./makeVitestConfig";
+export * from "./getVitestConfigIfExists";
+export * from "./createVitestConfig";
+export * from "./updateVitestConfig";
 export * from "./makeVitestSetup";
 export * from "./updateVitestTypes";
 export * from "./updateTsconfig";
