+++ packages/cli/src/helpers/updateVitestConfig/insertExportAssignmentFactory.test.ts
