+++ packages/cli/src/helpers/updateVitestConfig/insertExportAssignmentFactory.ts
