+++ packages/cli/src/helpers/updateVitestConfig/insertSetupFilesArrayFactory.test.ts
