+++ packages/cli/src/helpers/updateVitestConfig/insertSetupFilesArrayFactory.ts
