+++ packages/cli/src/helpers/updateVitestConfig/insertTestObjectFactory.test.ts
