+++ packages/cli/src/helpers/updateVitestConfig/insertVitestSetupFactory.test.ts
