+++ packages/cli/src/helpers/updateVitestConfig/insertVitestSetupFactory.ts
